refactor: Move authorization code to a new component
Move code related to authorization to a dedicated component to make
it available to other components.

The authorization service will be moved in a separate commit.
