+++ components/authorization/implementation/src/main/kotlin/permissions/OrganizationPermission.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.eclipse.apoapsis.ortserver.model.authorization
+package org.eclipse.apoapsis.ortserver.components.authorization.permissions
 
 import org.eclipse.apoapsis.ortserver.model.Organization
 import org.eclipse.apoapsis.ortserver.model.Product
