+++ components/authorization/implementation/src/main/kotlin/roles/OrganizationRole.kt
@@ -17,8 +17,9 @@
  * License-Filename: LICENSE
  */
 
-package org.eclipse.apoapsis.ortserver.model.authorization
+package org.eclipse.apoapsis.ortserver.components.authorization.roles
 
+import org.eclipse.apoapsis.ortserver.components.authorization.permissions.OrganizationPermission
 import org.eclipse.apoapsis.ortserver.model.Organization
 import org.eclipse.apoapsis.ortserver.model.Product
 import org.eclipse.apoapsis.ortserver.model.util.extractIdAfterPrefix
