+++ core/src/main/kotlin/api/AuthenticationException.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
+ * Copyright (C) 2022 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,19 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.eclipse.apoapsis.ortserver.model.authorization
+package org.eclipse.apoapsis.ortserver.core.api
 
-/**
- * Constants related to the superuser role that grants all permissions.
- */
-object Superuser {
-    /**
-     * The name of the superuser Keycloak role which grants all permissions.
-     */
-    const val ROLE_NAME = "superuser"
-
-    /**
-     * The name of the superuser Keycloak group which grants the [ROLE_NAME].
-     */
-    const val GROUP_NAME = "SUPERUSERS"
-}
+/** Used when the user does not have a valid authentication. */
+class AuthenticationException : RuntimeException()
