+++ core/src/test/kotlin/api/ErrorsIntegrationTest.kt
@@ -35,6 +35,7 @@ import org.eclipse.apoapsis.ortserver.clients.keycloak.test.KeycloakTestExtensio
 import org.eclipse.apoapsis.ortserver.clients.keycloak.test.TEST_SUBJECT_CLIENT
 import org.eclipse.apoapsis.ortserver.clients.keycloak.test.createKeycloakClientConfigurationForTestRealm
 import org.eclipse.apoapsis.ortserver.clients.keycloak.test.createKeycloakConfigMapForTestRealm
+import org.eclipse.apoapsis.ortserver.components.authorization.roles.Superuser
 import org.eclipse.apoapsis.ortserver.core.SUPERUSER
 import org.eclipse.apoapsis.ortserver.core.SUPERUSER_PASSWORD
 import org.eclipse.apoapsis.ortserver.core.createJsonClient
@@ -45,7 +46,6 @@ import org.eclipse.apoapsis.ortserver.core.setUpUserRoles
 import org.eclipse.apoapsis.ortserver.core.testutils.authNoDbConfig
 import org.eclipse.apoapsis.ortserver.core.testutils.ortServerTestApplication
 import org.eclipse.apoapsis.ortserver.dao.test.DatabaseTestExtension
-import org.eclipse.apoapsis.ortserver.model.authorization.Superuser
 import org.eclipse.apoapsis.ortserver.utils.test.Integration
 
 /**
