+++ settings.gradle.kts
@@ -28,6 +28,7 @@ include(":api:v1:mapping")
 include(":api:v1:model")
 include(":cli")
 include(":clients:keycloak")
+include(":components:authorization:implementation")
 include(":components:plugin-manager:api")
 include(":components:plugin-manager:implementation")
 include(":config")
